Added player + player movement
Player class, player movement added into the game.

player_data.gd is a Godot custom resource that allows us to add specific data and place it in the player scene in the inspector.

Added 2 test square images in the game for future test enemy/entity objects later on in the project